src/nts-business-verification/pyproject.toml
[project]
name = "data-go-mcp.nts-business-verification"
version = "0.1.0"
description = "MCP server for 국세청 사업자등록정보 진위확인 및 상태조회 서비스"
readme = "README.md"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
keywords = ["mcp", "korea", "api", "data.go.kr", "nts-business-verification"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.13.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[project.urls]
Homepage = "https://github.com/yourusername/data-go-mcp-servers"
Repository = "https://github.com/yourusername/data-go-mcp-servers"
Documentation = "https://github.com/yourusername/data-go-mcp-servers/tree/main/src/nts-business-verification"
"Bug Tracker" = "https://github.com/yourusername/data-go-mcp-servers/issues"
Changelog = "https://github.com/yourusername/data-go-mcp-servers/blob/main/src/nts-business-verification/CHANGELOG.md"

[project.scripts]
"data-go-mcp.nts-business-verification" = "data_go_mcp.nts_business_verification.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["data_go_mcp"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_default_fixture_loop_scope = "function"